Export public interfaces for Api and Contact
