+++ src/test/tsconfig.json
@@ -22,7 +22,7 @@
     "../lib/interfaces/io.ts",
     "../lib/login.ts",
     "../lib/request-io.ts",
-    "../lib/skype-http.ts",
+    "../lib/connect.ts",
     "../lib/utils.ts",
     "end-to-end/config.ts",
     "end-to-end/skype-http.spec.ts",
